+++ modules/containers_reflection/src/chrono.cpp
@@ -1,0 +1,5 @@
+//=================================================================================================
+// Copyright (C) 2023-2024 HEPHAESTUS Contributors
+//=================================================================================================
+
+#include "hephaestus/containers_reflection/chrono.h"
