+++ modules/format/CMakeLists.txt
@@ -20,7 +20,7 @@ set(SOURCES src/generic_formatter.cpp README.md include/hephaestus/format/generi
 # library target
 define_module_library(
   NAME format
-  PUBLIC_LINK_LIBS fmt::fmt hephaestus::utils reflectcpp::reflectcpp magic_enum::magic_enum
+  PUBLIC_LINK_LIBS fmt::fmt hephaestus::utils reflectcpp::reflectcpp magic_enum::magic_enum hephaestus::telemetry
   PRIVATE_LINK_LIBS ""
   SOURCES ${SOURCES}
   PUBLIC_INCLUDE_PATHS $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include> $<INSTALL_INTERFACE:include>
