+++ modules/format/tests/generic_formatter_tests.cpp
@@ -3,7 +3,7 @@
 //=================================================================================================
 
 #include <array>
-#include <chrono>
+#include <chrono>  // NOLINT(misc-include-cleaner)
 #include <cstdint>
 #include <iostream>
 #include <sstream>
@@ -90,46 +90,6 @@ TEST(GenericFormatterTests, TestFormatKnownObjectWithChronoDuration) {
   // Dummy test if general printers can be compiled
   std::cout << "cout: " << x << "\n" << std::flush;
   fmt::println("fmt: {}", x);
-}
-
-TEST(GenericFormatterTests, TestYAMLRoundtripWithChronoDuration) {
-  struct TestStruct {
-    std::string a = "test_value";
-    std::chrono::duration<double> b = std::chrono::minutes{ 42 };
-    std::chrono::milliseconds c = std::chrono::milliseconds{ 42 };
-  };
-  const TestStruct x{};
-  const auto yaml = rfl::yaml::write(x);
-  fmt::print("YAML: {}\n", yaml);
-
-  const auto parsed = rfl::yaml::read<TestStruct>(yaml);
-  EXPECT_TRUE(parsed.has_value());
-  EXPECT_EQ(parsed->a, x.a);
-  EXPECT_EQ(parsed->b, x.b);
-  EXPECT_EQ(parsed->c, x.c);
-}
-
-TEST(GenericFormatterTests, TestYAMLWithChronoDurationError) {
-  using namespace std::literals;
-  {
-    const std::string yaml = R"(100s)";
-    const auto parsed = rfl::yaml::read<std::chrono::seconds>(yaml);
-    EXPECT_TRUE(parsed.has_value());
-    EXPECT_EQ(*parsed, 100s);
-  }
-
-  {
-    const std::string yaml = R"(100)";
-    const auto parsed = rfl::yaml::read<std::chrono::seconds>(yaml);
-    EXPECT_TRUE(parsed.has_value());
-    EXPECT_EQ(*parsed, 100s);
-  }
-
-  // {
-  //   const std::string yaml = R"(100ms)";
-  //   const auto parsed = rfl::yaml::read<std::chrono::milliseconds>(yaml);
-  //   EXPECT_FALSE(parsed.has_value());
-  // }
 }
 
 TEST(GenericFormatterTests, TestFormatBounds) {
